test: split database-related tests into separate modules
